rag/agentic_rag/pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Zahir Shaikh",email = "zs4.machine.deep.learning@t-systems.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "openai (>=1.99.9,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "llama-index (>=0.13.2,<0.14.0)",
    "nemoguardrails (>=0.15.0,<0.16.0)",
    "flagembedding (>=1.3.5,<2.0.0)",
    "flagai (>=1.8.4,<2.0.0)",
    "llama-index-vector-stores-postgres (>=0.6.3,<0.7.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
